move RepresentationInterface to interfaces folder
