+++ cdisc_rules_engine/interfaces/__init__.py
@@ -1,10 +1,12 @@
 from .cache_service_interface import CacheServiceInterface
 from .data_reader_interface import DataReaderInterface
 from .factory_interface import FactoryInterface
+from .representation_interface import RepresentationInterface
 
 
 __all__ = [
     "CacheServiceInterface",
     "DataReaderInterface",
     "FactoryInterface",
+    "RepresentationInterface",
 ]
