+++ cdisc_rules_engine/interfaces/representation_interface.py
